% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationRules.R
\name{AssociationRules}
\alias{AssociationRules}
\title{Calculate Association rules with minimal support and confidence.}
\usage{
AssociationRules(Itemsets, FrequentItems, minsupport, minconfidence = 0,
  maxConsequentLength = 1)
}
\arguments{
\item{Itemsets}{Object of class TAMatrix that does contain the transactions underlying the rule
Mining.}

\item{FrequentItems}{Object of class FIMatrix that contains the frequent itemsets with minimal
support. The support used to calculate them must be the same as specified in minsupport. If no
dataset is provided the frequent Itemsets are calculated based on the Itemsets paramter.}

\item{minsupport}{Minimal support of the rules.}

\item{minconfidence}{Minimal confidence of the rules.}

\item{maxConsequentLength}{What should be the maximum length of the consequents of the rules?}
}
\value{
list with the rhs and lhs of the rules as sparse incident matrices. Also a vector
  containing the support and confidence of the inidivual rules is provided. The i'th element of
  the vector does represent the rule that is defined by the i'th column of the lhs and rhs
  matrix.
}
\description{
This function calculates association rules of consequent length >= 1 based on frequent itemsets.
If the Frequent itemsets are already calculated than one should provide them in the paramter
FrequentItems and set the paramter arefrequent to TRUE (default).
}
