% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{calcsupportFIMatrix}
\alias{calcsupportFIMatrix}
\title{Calculate support if missing.}
\usage{
calcsupportFIMatrix(input, support, dataset)
}
\arguments{
\item{input}{The frequent Itemsets as a ngTMatrix}

\item{support}{The support vector, if missing the support is calculated for the frequent itemsets
based on the transaction set in dataset.}

\item{dataset}{dataset containing the transaction data. May be of all classes that can be coerced
to TAMatrix by the function MakeTAMatrix.}
}
\value{
Vector of support values.
}
\description{
This function calculate the support for a frequent itemset with a transactions set if the 
support vector is missing and otherwise otherwise just return the support vector
}
