% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{colSums-FIMatrix}
\alias{colSums-FIMatrix}
\alias{colSums,FIMatrix-method}
\title{Give the colSums for the underlying matrix within an FIMatrix.}
\usage{
\S4method{colSums}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
numeric vector containing the sum of each column of the FIMatrix
}
\description{
In the matrix underlying the FIMatrix the rows represent the items and the columns to represent 
the itemsets. Here the sums of all columns should be calculated that are the number of items
for each itesmet.
}
