% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{colSums-Rules}
\alias{colSums-Rules}
\alias{colSums,Rules-method}
\title{Return the sum of each column for the for either the rhs or lhs of a rule}
\usage{
\S4method{colSums}{Rules}(x, lhs = TRUE)
}
\arguments{
\item{x}{Object of class Rules}

\item{lhs}{if true the sum of each column of the left-hand side are calculated, else the sum of 
each column of the right-hand side.}
}
\value{
numeric vector containing the sum of each column of either the rhs or the lhs.
}
\description{
Underlying a rules there is a right-hand side and a left-hand side. Both are stored as matrices. 
With this function one can calculated the sum of each column, that is the respective number
of items within each itemset for either the left-hand side or the right-hand side.
}
