% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{colSums-TAMatrix}
\alias{colSums-TAMatrix}
\alias{colSums,TAMatrix-method}
\title{Return the sum of each column for the underlying matrix within an TAMatrix}
\usage{
\S4method{colSums}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
numeric vector containing the sum of each column of the TAMatrix
}
\description{
In the matrix underlying the TAMatrix the rows represent the items and the columns  represent 
the itemsets. Here the sums of all columns should be calculated that are the number of items
for each itesmet.
}
