% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineCands.R
\name{CombineCands}
\alias{CombineCands}
\title{Combining incidence matrices of different length}
\usage{
CombineCands(list_input)
}
\arguments{
\item{list_input}{has to be a list of matrices. These matrices should have appropriate rownames describing the different items. The items should represent 
the different candidates.}
}
\description{
Combining incidence matrices of different length
}
