% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineRules.R
\name{CombineRules}
\alias{CombineRules}
\title{Combine a list of rules with potentially different items to one big matrix}
\usage{
CombineRules(inlist)
}
\arguments{
\item{inlist}{A list containing the different rule matrices. Theses should be incident matrices with the items as rows and the itemts as columns}
}
\value{
One big incident matrix including all rules.
}
\description{
Combine a list of rules with potentially different items to one big matrix
}
