% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{confidence-Rules}
\alias{confidence-Rules}
\alias{confidence,Rules-method}
\title{Extract confidence of all rules within a Rules object.}
\usage{
\S4method{confidence}{Rules}(object)
}
\arguments{
\item{object}{Object of class Rules}
}
\value{
Vector of confidence values from all Rules in x.
}
\description{
Extract confidence of all rules within a Rules object.
}
