% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{confidence}
\alias{confidence}
\title{S4 Generic to extract confidence vector from Rules object}
\usage{
confidence(object)
}
\arguments{
\item{input}{Object of Class Rules}
}
\value{
Vector of confidence values from all frequent itemsets
}
\description{
S4 Generic to extract confidence vector from Rules object
}
