% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{create_var_data}
\alias{create_var_data}
\title{Create a dataset with variable names, types, and
the corresponding terms related to that variable
in statistical models.}
\usage{
create_var_data(formulas, data)
}
\arguments{
\item{formulas}{a list of formulas}

\item{data}{the dataframe that will be used to fit models with \code{formulas}}
}
\description{
Create a dataset with variable names, types, and
the corresponding terms related to that variable
in statistical models.
}
