% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetRules.R
\name{DetRules}
\alias{DetRules}
\title{Determine all possible association rules of length len for given frequent transaction.}
\usage{
DetRules(Items, len, Items_support)
}
\arguments{
\item{Items}{Frequent Itemsets of different length, but at least two.}

\item{len}{length the resulting rules in the sense that on the right side there habe to be at least len items}
}
\description{
This function take the frequent transactions as input and generates rules of length len. It uses the Apriori-gen for more efficient generation
of rule candidates.
}
