% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetRules.R
\name{DetRules_1}
\alias{DetRules_1}
\title{Determine all possible association rules with consequent length 1for given frequent transaction.}
\usage{
DetRules_1(Items, Items_support)
}
\arguments{
\item{Items}{Frequent Itemsets of different length, but at least two.}

\item{Items_support}{Here the support of the frequent itemsets should be stored.}
}
\description{
This function take the frequent transactions as input and generates rules of length len. It uses the Apriori-gen for more efficient generation
of rule candidates.
}
