% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{DetRules_K}
\alias{DetRules_K}
\title{Determine all possible association rules with consequent length > 1 for given frequent itemset.}
\usage{
DetRules_K(rules)
}
\arguments{
\item{rules}{Object of class Rules containing the current rules with consequent length k}
}
\description{
This function takes the rules with consequent length k as input and generates rules of 
consequent length K + 1. For a rule having consequent length k it will return, if possible,
rules with consequent length  k + 1. It uses the Apriori-gen for more efficient generation 
of rule candidates.
}
