% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_data}
\alias{embed_data}
\title{embed a dataframe into a set of model specifications}
\usage{
embed_data(mspecs, data, ...)
}
\arguments{
\item{mspecs}{two-sided formula with outcome(s)
on the left hand side and exposure(s) on the
right hand side.}

\item{...}{key and value pairs specifying labels for variables in data.
For example, a variable called sbp may have the label of systolic blood
pressure. This could be set by writing sbp = 'systolic blood pressure'}
}
\description{
embed a dataframe into a set of model specifications
}
