% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_fits}
\alias{embed_fits}
\title{A function to fit various types of models based on apriori
specifications. This function should be applied within the
context of the \code{mutate()} function in the \code{dplyr} package.}
\usage{
embed_fits(object, engine, keep_models = FALSE, fill_models = TRUE,
  ...)
}
\arguments{
\item{object}{an object of class 'apriori'}

\item{engine}{a character value specifying the model type.
Valid options are
\itemize{
\item 'lm' for linear models
\item 'glm' for generalized linear models
\item 'gee' for generalized linear models fitted with
generalized estimating equations
\item 'cph' for Cox proportional hazards models.
}}

\item{keep_models}{(T/F) if TRUE, the model output will
be saved from its native R function. If FALSE, only
point estimates and covariance estimates will be returned.}

\item{...}{Arguments to be passed into the specified engine function.}

\item{family}{(character or function) a description of the error
distribution and link function to be used in the model. Valid options
for character values are 'binomial' and 'poisson' for binary and
count models, respectively. For regression models, family can be
left unspecified but engine should be set to 'lm'.}
}
\description{
A function to fit various types of models based on apriori
specifications. This function should be applied within the
context of the \code{mutate()} function in the \code{dplyr} package.
}
