% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_mspecs}
\alias{embed_mspecs}
\title{embed a set of model specifications into a primary hypothesis.}
\usage{
embed_mspecs(hypothesis, ..., verbose = FALSE)
}
\arguments{
\item{...}{Model specification objects created using \link{mspec_new},
\link{mspec_add}, \link{mspec_sub}, etc. (Control variables do belong
here).}
}
\description{
embed a set of model specifications into a primary hypothesis.
}
