% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Epub}
\alias{Epub}
\title{Download history of documents from the electronic publication platform}
\format{An object of class transaction with  15729  itemsets and 936 items. 
Session IDs and time stamps for transactions are also provided.}
\source{
http://epub.wu-wien.ac.at
}
\usage{
Epub
}
\description{
The \code{Epub} data set contains the download history of documents
from the electronic publication platform of the Vienna University of
Economics and Business Administration.
The data was recorded between Jan 2003 and Dec 2008.
}
\author{
Michael Hahsler
}
\keyword{datasets}
