% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{extract-Rules}
\alias{extract-Rules}
\alias{extract,Rules-method}
\title{Extract the FIMatrix from a rules object}
\usage{
\S4method{extract}{Rules}(object)
}
\arguments{
\item{object}{Object of class Rules}
}
\value{
FIMatrix containing the frequent itemsets based on which the rules where calculated.
}
\description{
Extract the FIMatrix from a rules object
}
