% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{extract}
\alias{extract}
\title{S4 Generic to extract frequent itemsets vector from Rules object}
\usage{
extract(object)
}
\arguments{
\item{input}{Object of Class Rules}
}
\value{
Object of Class FIMatrix
}
\description{
S4 Generic to extract frequent itemsets vector from Rules object
}
