% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFrequentSets.R
\name{ExtractFrequentSets}
\alias{ExtractFrequentSets}
\title{Extract the frequent itemset from output of arules apriori package}
\usage{
ExtractFrequentSets(Apriori_object)
}
\arguments{
\item{Apriori_object}{This has to be the ouput of the apriori command from the arules
package when looking for frequent itemsets}
}
\value{
Gives back a list with sets, the frequent itemsets and  their corresponding
 support. The format is similar to the output of ProjectApriori function FrequentItemsets.
}
\description{
This function does extract the sets and the support of a search for the frequent itemsets
 with arules apriori function
}
