% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractRules}
\alias{ExtractRules}
\title{Extract Rules}
\usage{
ExtractRules(rules, maxNumConsequent = NULL, order_by = NULL,
  decreasing = TRUE)
}
\arguments{
\item{rules}{An object of class Rules.}

\item{maxNumConsequent}{Maximal length the reported rule consequent should have.}

\item{order_by}{Specifiy up to four metrics out of support, confidence, lift, leverage by which
the given rules should be sorted.}

\item{decreasing}{Should the rules start with the smallest or highest values of the specified
metrics?}
}
\value{
The rules from the left hand and right hand side in the form of {It1, ... ItN} -> {ITK} 
in a data.frame. This data.frame does have columns lhs, rhs, unnamed, support and confidence.
}
\description{
This function extracts the rule from an object of  class Rules function in the form of 
{It1, ... ItN} -> {ITK}.
}
