% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{FIMatrix-class}
\alias{FIMatrix-class}
\title{S4 class to represent frequent itemsets in a sparse Matrix}
\description{
S4 class to represent frequent itemsets in a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{ngTMatrix containing the Frequent Itemset. Rows should represent the items and the 
columns should  represent the itemsets.}

\item{\code{support}}{Vector containing the support values for all Itemsets}
}}

