% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequentItemsets.R
\name{FindFrequentItemsets}
\alias{FindFrequentItemsets}
\title{Calculate frequent itemset with minimal support}
\usage{
FindFrequentItemsets(dataset, minsupport)
}
\arguments{
\item{dataset}{Transactions in an object of class TAMatrix, matrix, sparse matrix, data.frame or
transactions.}

\item{minsupport}{Minimal support value of the frequent itemsets}
}
\value{
Object of class FIMatrix conatining the frequent items
}
\description{
This function calculates the frequent itemset with a minimal support value.
}
\examples{
\donttest{
# Calculate the frequent itemsets with minimal support 0.03 
# and confidence 0.4 based on the dataset Groceries
Groceries_Fitems <- FindFrequentItemsets(Itemsets = Groceries, minsupport = 0.03)

# print the frequent itemsets
print(Groceries_Fitems)

# plot the frequent itemsets
plot(Groceries_Fitems)
}
}
