% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNDublicates.R
\name{FindNDuplicates}
\alias{FindNDuplicates}
\title{Give the number of duplicates of the matrix}
\usage{
FindNDuplicates(mat)
}
\arguments{
\item{mat}{Sparse matrix, where the rows are the items and the columns are the itemsets of type ngCmatrix}
}
\value{
The number of dupliates that occured in the matrix
}
\description{
Give the number of duplicates of the matrix
}
