% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_apri}
\alias{fit_apri}
\title{A wrapper function to fit various types of models based on apriori
specifications. This function should be applied within the
context of the \code{mutate()} function in the \code{dplyr} package.}
\usage{
fit_apri(formulas, data, engine = "lm", light_output = FALSE, ...)
}
\arguments{
\item{formulas}{a list of formula objects}

\item{data}{a dataframe containing the variables in the model.}

\item{engine}{a character value specifying the model type.
Valid options are
\itemize{
\item 'lm' for linear models
\item 'glm' for generalized linear models
\item 'gee' for generalized linear models fitted with
generalized estimating equations
\item 'cph' for Cox proportional hazards models.
}}

\item{light_output}{(T/F) if TRUE, the model output will
be abridged to a list containing regression point
estimates and covariance estimates. If FALSE, the full
model output will be returned.}

\item{...}{Arguments to be passed into the specified engine function.}

\item{family}{(character or function) a description of the error
distribution and link function to be used in the model. Valid options
for character values are 'binomial' and 'poisson' for binary and
count models, respectively. For regression models, family can be
left unspecified but engine should be set to 'lm'.}
}
\description{
A wrapper function to fit various types of models based on apriori
specifications. This function should be applied within the
context of the \code{mutate()} function in the \code{dplyr} package.
}
