% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_format.R
\name{fmt_effect}
\alias{fmt_effect}
\title{format estimated regression coefficients from a model.
It is common to present effects in terms of odds ratios,
prevalence ratios, or hazard ratios by applying the exponential
function (\code{exp}) to the regression coefficients. It is also
common to present these ratios with 95% confidence intervals
(i.e., \code{conf_level = 0.95}) in model summary tables.}
\usage{
fmt_effect(effect, error, transform = NULL, conf_level = NULL,
  reference_index = NULL, reference_label = "1 (reference)")
}
\arguments{
\item{effect}{a numeric vector of estimated regression coefficients}

\item{error}{a numeric vector of standard error estimates corresponding
to the regression coefficients in the \code{effect} vector.}

\item{transform}{a function that will be applied to the values in \code{effect}
(e.g., exponentiating to produce odds ratios).}

\item{conf_level}{(optional) the desired confidence level for interval
estimation. It is common to use \code{conf_level = 0.95} to produce 95%
confidence intervals.}

\item{reference_index}{(optional) If there is a referent group in the
\code{effect} values, specify the indices of the group. For example, if
\code{effect = c(0, 1/2, 1)} describing the relative increase in hazard
for groups B and C relative to A, then \code{reference_index = 1}.}

\item{reference_label}{(optional) a string indicating how the reference
group will be labeled in representations of the model effects.}
}
\description{
format estimated regression coefficients from a model.
It is common to present effects in terms of odds ratios,
prevalence ratios, or hazard ratios by applying the exponential
function (\code{exp}) to the regression coefficients. It is also
common to present these ratios with 95% confidence intervals
(i.e., \code{conf_level = 0.95}) in model summary tables.
}
