% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{FIMatrix}
\alias{FIMatrix}
\title{An S4 class to represent an frequent Itemsets in a sparse Matrix}
\description{
An S4 class to represent an frequent Itemsets in a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{Frequent Itemset Matrix in sparse representation}

\item{\code{support}}{Corresponding support}
}}

