% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCandidates.R
\name{GenCandidates}
\alias{GenCandidates}
\title{Apriori Gen of candidates K-1}
\usage{
GenCandidates(L)
}
\arguments{
\item{L}{Incident matrix of the transactions}
}
\value{
The candidate for the next iteration of Apriori
}
\description{
Takes in a incident matrix of transaction, where the rows do describe the potential items and each colomn does represent the itemsets.
}
