% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_model_variables}
\alias{get_model_variables}
\title{Create a dataset with variable names, types, and
the corresponding terms related to that variable
in statistical models.}
\usage{
get_model_variables(formulas, data)
}
\arguments{
\item{formulas}{a list of formulas}

\item{data}{the dataframe that will be used to fit models with \code{formulas}}
}
\description{
Create a dataset with variable names, types, and
the corresponding terms related to that variable
in statistical models.
}
