% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiveUniqueCol.R
\name{GiveUniqueCol}
\alias{GiveUniqueCol}
\title{Give only the unique columns of a sparse matrix.}
\usage{
GiveUniqueCol(mat)
}
\arguments{
\item{mat}{Sparse matrix, where the rows are the items and the columns are the itemsets of type ngCmatrix}
}
\value{
Returns only non-duplicate columns of matrix. Ordering might be changed.
}
\description{
Give only the unique columns of a sparse matrix.
}
