% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{hoist_effect}
\alias{hoist_effect}
\title{get the estimated effects for a designated exposure variable
from a set of statistical models fitted using apriori specs.}
\usage{
hoist_effect(data, models, effect, ci = NULL, transform = NULL,
  reference_label = "1 (reference)")
}
\arguments{
\item{data}{a dataset with containing a list of models}

\item{models}{a column in dataset containing a list of models}

\item{effect}{(character or literal) the effect to hoist}

\item{ci}{a numeric value between 0 and 1 indicating desired
confidence level for interval estimates. If left unspecified,
confidence intervals will not be included in the output.}

\item{fun}{a function that will be applied to the model's
estimated regression coefficients. For example, it is common
to exponentiate regression coefficients from logistic and
proportional hazards models to present odds and hazard ratios,
respectively.}
}
\description{
get the estimated effects for a designated exposure variable
from a set of statistical models fitted using apriori specs.
}
