% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{hypothesize_that}
\alias{hypothesize_that}
\title{initialize an a priori analysis with a hypothesis.}
\usage{
hypothesize_that(formula)
}
\arguments{
\item{formula}{a two-sided formula. Outcomes go on the
left hand side of the formula, and exposures on the right.}
}
\description{
initialize an a priori analysis with a hypothesis.
}
\note{
Control variables do not belong here. Instead, they
are specified in the \code{\link{mspec_new}},
\code{\link{mspec_add}}, \code{\link{mspec_rmv}}, and
\code{\link{mspec_sub}} functions.
}
