% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{items-FIMatrix}
\alias{items-FIMatrix}
\alias{items,FIMatrix-method}
\title{Export the item names for a FIMatrix}
\usage{
\S4method{items}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
Names of all items in FIMatrix.
}
\description{
Export the item names for a FIMatrix
}
