% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{items-Rules}
\alias{items-Rules}
\alias{items,Rules-method}
\title{Export the item names for a Rules object.}
\usage{
\S4method{items}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
Vector containing the names of all items in Rules.
}
\description{
Export the item names for a Rules object.
}
