% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{items-TAMatrix}
\alias{items-TAMatrix}
\alias{items,TAMatrix-method}
\title{Export the item names for a TAMatrix.}
\usage{
\S4method{items}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
Names of all items in TAMatrix.
}
\description{
Export the item names for a TAMatrix.
}
