% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{items}
\alias{items}
\title{Extracing the names of the items.}
\usage{
items(x)
}
\arguments{
\item{x}{Object to extract the items from}
}
\value{
A character vector containing the names of the items
}
\description{
Extracing the names of the items.
}
