% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{length-FIMatrix}
\alias{length-FIMatrix}
\alias{length,FIMatrix-method}
\title{Determine the number of items in a FIMatrix}
\usage{
\S4method{length}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
Number of frequent itemsets within the FIMatrix
}
\description{
The length function for the FIMatrix returns the number of frequent items.
}
