% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{length-Rules}
\alias{length-Rules}
\alias{length,Rules-method}
\title{Determine the number of rules in a Rules object}
\usage{
\S4method{length}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
Number of Rules in x.
}
\description{
The length function for the Rules class returns the number of rules.
}
