% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{length-TAMatrix}
\alias{length-TAMatrix}
\alias{length,TAMatrix-method}
\title{Determine the number of items in a TAMatrix}
\usage{
\S4method{length}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
Number of  items within the TAMatrix
}
\description{
The length function for the TAMatrix returns the number of items in that TAMatrix.
}
