% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{leverage-Rules}
\alias{leverage-Rules}
\alias{leverage,Rules-method}
\title{Extract leverage of all rules within a Rules object.}
\usage{
\S4method{leverage}{Rules}(object)
}
\arguments{
\item{object}{Object of class Rules}
}
\value{
Vector of leverage values from all Rules in x.
}
\description{
Extract leverage of all rules within a Rules object.
}
