% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{leverage}
\alias{leverage}
\title{S4 Generic to extract leverage vector from Rules object}
\usage{
leverage(object)
}
\arguments{
\item{input}{Object of Class Rules}
}
\value{
Vector of leverage values from all frequent itemsets
}
\description{
S4 Generic to extract leverage vector from Rules object
}
