% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{lift-Rules}
\alias{lift-Rules}
\alias{lift,Rules-method}
\title{Extract lift of all rules within a Rules object.}
\usage{
\S4method{lift}{Rules}(object)
}
\arguments{
\item{object}{Object of class Rules}
}
\value{
Vector of lift values from all Rules in x.
}
\description{
Extract lift of all rules within a Rules object.
}
