% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{lift}
\alias{lift}
\title{S4 Generic to extract lift vector from Rules object}
\usage{
lift(object)
}
\arguments{
\item{input}{Object of Class Rules}
}
\value{
Vector of lift values from all frequent itemsets
}
\description{
S4 Generic to extract lift vector from Rules object
}
