% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeFIMatrix-transactions}
\alias{makeFIMatrix-transactions}
\alias{makeFIMatrix.transactions}
\title{Function to create a FIMatrix from an transactions}
\usage{
makeFIMatrix.transactions(input, support, dataset)
}
\arguments{
\item{input}{Object of class transactions}

\item{support}{Support values for the transactions}

\item{dataset}{Underlying transactions. Has to be supplied if the support vector is missing and
has to be calculated.}
}
\value{
Object of class FIMatirx containing correct support values
}
\description{
Function to create a FIMatrix from an transactions
}
