% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeFIMatrix.FIMatrix}
\alias{makeFIMatrix.FIMatrix}
\title{Function to create a FIMatrix from an FIMatrix}
\usage{
makeFIMatrix.FIMatrix(input, support, dataset)
}
\arguments{
\item{input}{Object of class FIMatrix}

\item{support}{Support values for the FIMatrix}

\item{dataset}{Underlying transactions. Has to be supplied if the support vector is missing.}
}
\value{
Object of class FIMatrix containing the same itemset as the input.
}
\description{
Function to create a FIMatrix from an FIMatrix
}
