% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeFIMatrix.lgTMatrix}
\alias{makeFIMatrix.lgTMatrix}
\title{Function to create a FIMatrix from an lgTMatrix}
\usage{
makeFIMatrix.lgTMatrix(input, support, dataset)
}
\arguments{
\item{input}{Object of class lgTMatrix}

\item{support}{Support values for the lgTMatrix}

\item{dataset}{Underlying data.frame Has to be supplied if the support vector is missing and
has to be calculated.}
}
\value{
Object of class FIMatirx containing correct support values
}
\description{
Function to create a FIMatrix from an lgTMatrix
}
