% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeFIMatrix.matrix}
\alias{makeFIMatrix.matrix}
\title{Function to create a FIMatrix from an matrix}
\usage{
makeFIMatrix.matrix(input, support, dataset)
}
\arguments{
\item{input}{Object of class matrix}

\item{support}{Support values for the matrix}

\item{dataset}{Underlying matrix Has to be supplied if the support vector is missing and
has to be calculated.}
}
\value{
Object of class FIMatirx containing correct support values
}
\description{
Function to create a FIMatrix from an matrix
}
