% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeFIMatrix.ngTMatrix}
\alias{makeFIMatrix.ngTMatrix}
\title{Function to create a FIMatrix from an ngTMatrix}
\usage{
makeFIMatrix.ngTMatrix(input, support, dataset)
}
\arguments{
\item{input}{Object of class ngTMatrix}

\item{support}{Support values for the ngTMatrix}

\item{dataset}{Underlying data.frame Has to be supplied if the support vector is missing and
has to be calculated.}
}
\value{
Object of class FIMatirx containing correct support values
}
\description{
Function to create a FIMatrix from an ngTMatrix
}
