% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeFIMatrix}
\alias{makeFIMatrix}
\title{Take different input and make them a FIMatrix}
\usage{
makeFIMatrix(input, support, dataset, minsupport)
}
\arguments{
\item{mat}{Object of class matrix, sparse matrix or Formel class transaction (arules
class) containing the frequent itemsets.}
}
\value{
The non-sparse matrix
}
\description{
Takes a object of class Matrix, ngCMatrix, ngTMatrix or formel class transaction and return a
corresponding object of class FIMatrix. The rows should represent the itemset and be named.
The columns should describe
}
