% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.default}
\alias{makeTAMatrix.default}
\title{Default method for not implemented classes}
\usage{
makeTAMatrix.default(input)
}
\arguments{
\item{input}{Object of uknown class}
}
\value{
Error message
}
\description{
Default method for not implemented classes
}
