% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.lgCMatrix}
\alias{makeTAMatrix.lgCMatrix}
\title{Function to create a TAMatrix from an object of class lgCMatrix}
\usage{
makeTAMatrix.lgCMatrix(input)
}
\arguments{
\item{input}{Object of class lgCMatrix}
}
\value{
Object of class data.frame containing the same itemset as the input.
}
\description{
Function to create a TAMatrix from an object of class lgCMatrix
}
