% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.lgTMatrix}
\alias{makeTAMatrix.lgTMatrix}
\title{Function to create a TAMatrix from an object of class lgTMatrix}
\usage{
makeTAMatrix.lgTMatrix(input)
}
\arguments{
\item{input}{Object of class lgTMatrix}
}
\value{
Object of class data.frame containing the same itemset as the input.
}
\description{
Function to create a TAMatrix from an object of class lgTMatrix
}
