% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.matrix}
\alias{makeTAMatrix.matrix}
\title{Function to create a TAMatrix from an object of class matrix}
\usage{
makeTAMatrix.matrix(input)
}
\arguments{
\item{input}{Object of class matrix}
}
\value{
Object of class matrix containing the same itemset as the input.
}
\description{
Function to create a TAMatrix from an object of class matrix
}
