% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.ngCMatrix}
\alias{makeTAMatrix.ngCMatrix}
\title{Function to create a TAMatrix from an object of class ngCMatrix}
\usage{
makeTAMatrix.ngCMatrix(input)
}
\arguments{
\item{input}{Object of class ngCMatrix}
}
\value{
Object of class data.frame containing the same itemset as the input.
}
\description{
Function to create a TAMatrix from an object of class ngCMatrix
}
