% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.ngTMatrix}
\alias{makeTAMatrix.ngTMatrix}
\title{Function to create a TAMatrix from an object of class ngTMatrix}
\usage{
makeTAMatrix.ngTMatrix(input)
}
\arguments{
\item{input}{Object of class ngTMatrix}
}
\value{
Object of class data.frame containing the same itemset as the input.
}
\description{
Function to create a TAMatrix from an object of class ngTMatrix
}
