% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.TAMatrix}
\alias{makeTAMatrix.TAMatrix}
\title{Function to create a TAMatrix from an TAMatrix}
\usage{
makeTAMatrix.TAMatrix(input)
}
\arguments{
\item{input}{Object of class TAMatrix}
}
\value{
Object of class TAMatrix containing the same itemset as the input.
}
\description{
Function to create a TAMatrix from an TAMatrix
}
