% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{makeTAMatrix.transactions}
\alias{makeTAMatrix.transactions}
\title{Function to create a TAMatrix from an object of class transaction (coming from the arules 
package)}
\usage{
makeTAMatrix.transactions(input)
}
\arguments{
\item{input}{Object of class transactions}
}
\value{
Object of class transactions containing the same itemset as the input
}
\description{
Function to create a TAMatrix from an object of class transaction (coming from the arules 
package)
}
