% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTansactionMatrix.R
\name{makeTansactionMatrix}
\alias{makeTansactionMatrix}
\title{Make Transaction Dataset from arules to non- sparse matrix}
\usage{
makeTansactionMatrix(mat)
}
\arguments{
\item{mat}{Incident matrix of the transactions}
}
\value{
The non-sparse matrix
}
\description{
Takes a object of class ... from arules and makes non-sparse matrix from it.
}
