% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransactionMatrix.R
\name{makeTransactionMatrix}
\alias{makeTransactionMatrix}
\title{Make Transaction Dataset from arules to non- sparse matrix}
\usage{
makeTransactionMatrix(mat)
}
\arguments{
\item{mat}{Incident matrix of the transactions}
}
\value{
The non-sparse matrix
}
\description{
Takes a object of class ... from arules and makes non-sparse matrix from it.
}
