% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{mspec_push}
\alias{mspec_push}
\title{push a model specification into a model formula.}
\usage{
mspec_push(mspec, formula)
}
\arguments{
\item{mspec}{a model specification object}

\item{formula}{a formula with outcome and exposure.}
}
\description{
push a model specification into a model formula.
}
