% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{ncol-FIMatrix}
\alias{ncol-FIMatrix}
\alias{ncol,FIMatrix-method}
\title{Give the number of columns of underlying matrix in an FIMatrix.}
\usage{
\S4method{ncol}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
number of columns / itemsets in the FIMatrix
}
\description{
This number does represent the number of itemsets within that FIMatrix
}
