% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{ncol-Rules}
\alias{ncol-Rules}
\alias{ncol,Rules-method}
\title{Return the number of columns of underlying matrix in an Rules object}
\usage{
\S4method{ncol}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
number of columns / Rules in the Rules object.
}
\description{
Although a Rules object does have left-hand side and a right hand-side the number of columns for
both does represent the number rules and therefore should be the same for both sides. This functions
simply uses the left-hand sides as proxy.
}
