% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{ncol-TAMatrix}
\alias{ncol-TAMatrix}
\alias{ncol,TAMatrix-method}
\title{Give the number of columns of underlying matrix in an TAMatrix.}
\usage{
\S4method{ncol}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
number of columns / itemsets in the TAMatrix
}
\description{
This number does represent the number of itemsets within that TAMatrix
}
