% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{nrow-FIMatrix}
\alias{nrow-FIMatrix}
\alias{nrow,FIMatrix-method}
\title{Give the number of rows of underlying matrix in an FIMatrix.}
\usage{
\S4method{nrow}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
number of rows / items in the FIMatrix
}
\description{
This number does represent the number of items within that FIMatrix
}
