% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{nrow-Rules}
\alias{nrow-Rules}
\alias{nrow,Rules-method}
\title{Give the number of rows of underlying matrix in an Rules object.}
\usage{
\S4method{nrow}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
number of rows / total number of possible items in the Rules object.
}
\description{
Although a Rules object does have left-hand side and a right hand-side the number of rows for
both does represent the number items and therefore should be the same for both sides. This functions
simply uses the left-hand sides as proxy.
}
