% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{nrow-TAMatrix}
\alias{nrow-TAMatrix}
\alias{nrow,TAMatrix-method}
\title{Return the number of rows of underlying matrix in an TAMatrix.}
\usage{
\S4method{nrow}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
number of rows / items in the TAMatrix
}
\description{
This number does represent the number of items within that TAMatrix
}
