% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceTypes.R
\name{outputTAMatrix}
\alias{outputTAMatrix}
\title{Function to create a TAMatrix from an ngTMatrix.}
\usage{
outputTAMatrix(out_mat)
}
\arguments{
\item{out_mat}{Object of class ngTMatrix}
}
\value{
Object of class TAMatrix corresponding to the ngTMatrix
}
\description{
Function to create a TAMatrix from an ngTMatrix.
}
