% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{plot-Rules}
\alias{plot-Rules}
\alias{plot,Rules-method}
\title{Plot an Rules object.}
\usage{
\S4method{plot}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
Scatter plot of support versus confidence with lift as a color gradient
}
\description{
The plot function gives a scatter plot with the support on the x-axis, the confidence on the y-axis
and the lift as a color gradient.
}
