% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{plot-TAMatrix}
\alias{plot-TAMatrix}
\alias{plot,TAMatrix-method}
\title{Plot an TAMatrix object.}
\usage{
\S4method{plot}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
histogram of the length of the itemsets within the TA Matrix.
}
\description{
The plot function for TA Matrix does give a histgram of the itemset length.
}
