% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{print-FIMatrix}
\alias{print-FIMatrix}
\alias{print,FIMatrix-method}
\title{Definition of show method for FIMatrix}
\usage{
\S4method{print}{FIMatrix}(x, descending = TRUE)
}
\arguments{
\item{x}{Object of class FIMatrix}

\item{descending}{Logical value per default TRUE.}
}
\value{
The frequent itemsets ordererd by their support
}
\description{
The print function prints out all the frequent itemsets sorted by their support value.
}
