% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{print-Rules}
\alias{print-Rules}
\alias{print,Rules-method}
\title{Definition of show method for Rules}
\usage{
\S4method{print}{Rules}(x, maxNumConsequent = 1, order_by = NULL,
  decreasing = TRUE)
}
\arguments{
\item{x}{Object of class Rules}

\item{maxNumConsequent}{The maximum length of consequents that the rules of the ouput should 
have. In Default all rules are shown.}

\item{order_by}{Specifiy up to four metrics out of support, confidence, lift, leverage by which
the given rules should be sorted. The first one used first and son on.}

\item{decreasing}{Should the rules start with the smallest or highest values of the specified
metrics?}
}
\value{
The rules from the left hand and right hand side in the form of {It1, ... ItN} -> {ITK} 
in a data.frame. This data.frame does have columns lhs, rhs, unnamed, support and confidence.
}
\description{
The print function prints out all the rules in the object sorted by specified matrices.
}
