% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{print-TAMatrix}
\alias{print-TAMatrix}
\alias{print,TAMatrix-method}
\title{Definition of print method for TAMatrix}
\usage{
\S4method{print}{TAMatrix}(x, descending = TRUE)
}
\arguments{
\item{x}{Object of class TAMatrix}

\item{descending}{Starting with the highest or lowest count?}
}
\value{
The frequent itemsets ordererd by their occurence
}
\description{
The print function prints out all items of the TAMatrix with their respective counts. There are
sorted descending by these counts
}
