% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.apri_mspec}
\alias{print.apri_mspec}
\title{print an apriori analysis plan}
\usage{
\method{print}{apri_mspec}(x, ...)
}
\arguments{
\item{x}{An a-priori object to be printed}

\item{...}{additional arguments passed to print}
}
\description{
print an apriori analysis plan
}
