% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{print.apri_spec}
\alias{print.apri_spec}
\title{print a model specification}
\usage{
\method{print}{apri_spec}(x, ...)
}
\arguments{
\item{x}{An a-priori model specification object to be printed}

\item{...}{additional arguments passed to print}
}
\description{
print a model specification
}
