% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.apriori}
\alias{print.apriori}
\title{print a model specification}
\usage{
\method{print}{apriori}(x, ...)
}
\arguments{
\item{x}{An a-priori model specification object to be printed}

\item{...}{additional arguments passed to print}
}
\description{
print a model specification
}
