% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{prune-FIMatrix}
\alias{prune-FIMatrix}
\alias{prune,FIMatrix-method}
\title{Prune method for objects of class FIMatrix}
\usage{
\S4method{prune}{FIMatrix}(object, Support)
}
\arguments{
\item{object}{Object of class FIMatrix.}

\item{Support}{Minimal support the pruned FIMatrix should have.}
}
\value{
Pruned object of class FIMatrix
}
\description{
With this function one can delete all itemsets from an FIMatrix that do not have minimal support.
}
