% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{prune-Rules}
\alias{prune-Rules}
\alias{prune,Rules-method}
\title{Prune method for objects of class Rules}
\usage{
\S4method{prune}{Rules}(object, Support, Confidence, Lift, Leverage,
  inv_Lift = FALSE, inv_lev = FALSE)
}
\arguments{
\item{object}{Object of class Rules}

\item{Support}{Minimal support the output rules should have.}

\item{Confidence}{Minimal confidence the output rules should have.}

\item{Lift}{Minimal Lift the output rules should have.}

\item{Leverage}{Minimal Leverage the output rules should have.}

\item{inv_Lift}{Pruning based on minimal or maximal lift?}

\item{inv_Leverage}{Pruning based on minimal or maximal leverage?}
}
\value{
Pruned object of class Rules
}
\description{
With this function one can delete all itemsets from an Rules that do not have minimal support,
confidence lift or leverage.
}
