% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apriori_methods.R
\name{pull_analysis}
\alias{pull_analysis}
\title{extract the analysis component from an apriori object.
This makes it easier to pass apriori objects into other
functions for presentation.}
\usage{
pull_analysis(object, ...)
}
\arguments{
\item{object}{an apriori object}

\item{...}{further arguments passed on to \code{\link[dplyr]{select}}}
}
\description{
extract the analysis component from an apriori object.
This makes it easier to pass apriori objects into other
functions for presentation.
}
