% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{qplot-FIMatrix}
\alias{qplot-FIMatrix}
\alias{qplot,FIMatrix-method}
\title{Plot an FI-matrices with ggplot}
\usage{
\S4method{qplot}{FIMatrix}(x, col = "red", alpha = 0.1, type = c("hist",
  "scatter"))
}
\arguments{
\item{x}{Object of class FIMatrix}

\item{col}{colour that is given to plot function.}

\item{alpha}{Alpha value value that is given to qplot}

\item{type}{should it be a scatter plot or a histogram?}
}
\value{
Scatter plot of Itemsize vs support.
}
\description{
The qplot function gives a scatter plot with the support on the y-axis and the itemset-length on
the x-axis
}
