% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{qplot-Rules}
\alias{qplot-Rules}
\alias{qplot,Rules-method}
\title{Plot a Rules object wiht ggplot2}
\usage{
\S4method{qplot}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
Scatter qplot of support versus confidence, lift as a color gradient..
}
\description{
The plot function gives a scatter plot with the support on the x-axis, the confidence on the y-axis
and the lift as a color gradient.
}
