% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{qplot-TAMatrix}
\alias{qplot-TAMatrix}
\alias{qplot,TAMatrix-method}
\title{Plot an TAMatrix object with ggplot.}
\usage{
\S4method{qplot}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\value{
histogram of the length of the itemsets within the TA Matrix.
}
\description{
The qplot function for TA Matrix does give a histgram of the itemset length.
}
