% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{qplot}
\alias{qplot}
\title{ggplot plot methods for Rules, TAMatrix and FIMatrix}
\usage{
qplot(x, ...)
}
\arguments{
\item{x}{object of class Rules, TAMatrix or FIMatrix}

\item{...}{Further information that should be passed down to plot function}
}
\value{
Plot
}
\description{
ggplot plot methods for Rules, TAMatrix and FIMatrix
}
