% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{rowSums-FIMatrix}
\alias{rowSums-FIMatrix}
\alias{rowSums,FIMatrix-method}
\title{Return the row Sums for the underlying matrix within an FIMatrix}
\usage{
\S4method{rowSums}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
numeric vector containing the sum of each row of the FIMatrix
}
\description{
In the matrix underlying the FIMatrix the rows represent the items and the columns to represent 
the itemsets. Here the sums of each row should be calculated that are the number of occurences
of each item.
}
