% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpriori.R
\docType{package}
\name{rpriori-package}
\alias{rpriori}
\alias{rpriori-package}
\title{rpriori: A grammar for testing apriori hypotheses in R.}
\description{
To learn more about rpriori, start with the vignettes:
\code{browseVignettes(package = "rpriori")}
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu}

}
