% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Rules-class}
\alias{Rules-class}
\title{S4 class to store association rules and relevant quality metrics}
\description{
S4 class to store association rules and relevant quality metrics
}
\section{Slots}{

\describe{
\item{\code{lhs}}{ngTMatrix describing the lhs of the rules.}

\item{\code{rhs}}{ngTMatrix describing the rhs of the rules.}

\item{\code{support}}{Support of  the Rules.}

\item{\code{confidence}}{Confidence of Rules.}

\item{\code{lift}}{Lift of the rules.}

\item{\code{leverage}}{Leverage of the rules.}

\item{\code{itemsetID}}{This represent a vector of unique identifiers for the different frequent
itemsets the rules stem from. Used for internal computation.}

\item{\code{FrequentItemsets}}{Frequent Itemsets as an FIMatrix object containing the frequent items 
the rules were created from.}
}}

