% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{select-Rules}
\alias{select-Rules}
\alias{select,Rules-method}
\title{Subsetting of an Rules object.}
\usage{
\S4method{select}{Rules}(x, i, j)
}
\arguments{
\item{x}{Object of class Rules}

\item{i}{Either the rows represented by their row number or a logical vector of length number of 
row of Rules. If i is missing or NULL all rows are selected.}

\item{j}{Either the columns represented by their columns numbers or logical vector of length 
number of columns in Rules. If j is missing or NULL all columns are selected.}
}
\value{
subsetted Rules object.
}
\description{
An Rules does contain the matrix of itemsets as well as the a vectors that contains the support,
confidence, lift and leverage for all rules.Therefore, both are logically connected and 
when a Rules is subsetted column- wise the other vectors are subsetted as well.
}
