% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{select-TAMatrix}
\alias{select-TAMatrix}
\alias{select,TAMatrix-method}
\title{Subsetting of an TAMatrix}
\usage{
\S4method{select}{TAMatrix}(x, i, j)
}
\arguments{
\item{x}{Object of class TAMatrix}

\item{i}{Either the rows represented by their row number or a logical vector of length number of 
row of TAMatrix.}

\item{j}{Either the columns represented by their columns numbers or logical vector of length 
number of columns in TAMatrix}
}
\value{
subsetted TAMatrix
}
\description{
An TAMatrix does contain the matrix of all transactions as well as the dimensions of that 
matrix and the names of all items. Therefore, all these parts are logically connected and have
to be changed when the matrix is subsetted.
}
