% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{select}
\alias{select}
\title{Subsetting  of TAMatrix, FIMatrix and Rules class.}
\usage{
select(x, i, j)
}
\arguments{
\item{x}{Object to subset}

\item{i}{Either rows represented by their row number or logical vector of length
number of rows}

\item{j}{Either columns represented by their column number or logical vector of length
number of columns}
}
\value{
A character vector containing the names of the items
}
\description{
Subsetting  of TAMatrix, FIMatrix and Rules class.
}
