% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{sf_labels}
\alias{sf_labels}
\title{helper function to work with survival formulas.
Generates labels that look reasonable in printed dataframes}
\usage{
sf_labels(string)
}
\arguments{
\item{string}{the left hand side of a survival formula.}
}
\description{
helper function to work with survival formulas.
Generates labels that look reasonable in printed dataframes
}
