% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{show-FIMatrix}
\alias{show-FIMatrix}
\alias{show,FIMatrix-method}
\title{Definition of show method for FIMatrix}
\usage{
\S4method{show}{FIMatrix}(object)
}
\arguments{
\item{object}{Object of class FIMatrix}
}
\value{
Short message stating the number of frequent itemsets.
}
\description{
The show function prints out the number of frequent itemsets in the FIMatrix.
}
