% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{show-Rules}
\alias{show-Rules}
\alias{show,Rules-method}
\title{Definition of show method for Rules}
\usage{
\S4method{show}{Rules}(object)
}
\arguments{
\item{object}{Object of class Rules}
}
\value{
Short message stating the number of Rules.
}
\description{
The show function prints out the number of Rules
}
