% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{show-TAMatrix}
\alias{show-TAMatrix}
\alias{show,TAMatrix-method}
\title{Definition of show method for TAMatrix}
\usage{
\S4method{show}{TAMatrix}(object)
}
\arguments{
\item{object}{Object of class TAMatrix}
}
\value{
Short message stating the number of items.
}
\description{
The show function prints out the number of items in the TAMatrix
}
