% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_embed}
\alias{spec_embed}
\title{combine all pairs of outcome ~ exposure with
all model specifications in a given set.}
\usage{
spec_embed(formula, ...)
}
\arguments{
\item{formula}{two-sided formula with outcome(s)
on the left hand side and exposure(s) on the
right hand side.}

\item{...}{model specification objects.}
}
\description{
combine all pairs of outcome ~ exposure with
all model specifications in a given set.
}
