% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_push}
\alias{spec_push}
\title{push a model specification into a model formula.}
\usage{
spec_push(model_spec, formula)
}
\arguments{
\item{model_spec}{a model specification object}

\item{formula}{a formula with outcome and exposure.}
}
\description{
push a model specification into a model formula.
}
