% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{summary-FIMatrix}
\alias{summary-FIMatrix}
\alias{summary,FIMatrix-method}
\title{Summary for FI-matrices.}
\usage{
\S4method{summary}{FIMatrix}(object)
}
\arguments{
\item{object}{Object of class FIMatrix}
}
\value{
Summary information about the FImatrix
}
\description{
The sumamary function gives general information about the frequent itemsets.
}
