% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{summary-Rules}
\alias{summary-Rules}
\alias{summary,Rules-method}
\title{Summary funtion for Rules object}
\usage{
\S4method{summary}{Rules}(object)
}
\arguments{
\item{object}{Object of class Rules}
}
\value{
Summary information about the Rules
}
\description{
The summary for the rules object does give some general information on the quality of the rules.
}
