% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{summary-TAMatrix}
\alias{summary-TAMatrix}
\alias{summary,TAMatrix-method}
\title{Summary for TAMatrix objects}
\usage{
\S4method{summary}{TAMatrix}(object)
}
\arguments{
\item{object}{Object of class TAMatrix}
}
\value{
Summary information about the TAMatrix
}
\description{
The summary function gives general information about the TAMatrix such as the density or the 
distribution of the length of the itemsets
}
