% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.apri_fit}
\alias{summary.apri_fit}
\title{summarize an \code{apri_fit} object.}
\usage{
\method{summary}{apri_fit}(object)
}
\arguments{
\item{object}{an object of class \code{apri_fit}}
}
\description{
summarize an \code{apri_fit} object.
}
