% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{support-FIMatrix}
\alias{support-FIMatrix}
\alias{support,FIMatrix-method}
\title{Extract the support of itemsets in class FIMatrix}
\usage{
\S4method{support}{FIMatrix}(object)
}
\arguments{
\item{object}{Object of class FIMatrix}
}
\value{
A numeric vector containing the support values of the itemsets in the FIMatrix.
}
\description{
Extract the support of itemsets in class FIMatrix
}
