% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{support-Rules}
\alias{support-Rules}
\alias{support,Rules-method}
\title{Extract the support of itemsets in class Rules}
\usage{
\S4method{support}{Rules}(x)
}
\arguments{
\item{x}{Object of class Rules}
}
\value{
A numeric vector containing the support values of Rules.
}
\description{
Extract the support of itemsets in class Rules
}
