% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{support}
\alias{support}
\title{Extracing the support of objects}
\usage{
support(object)
}
\arguments{
\item{support}{Object to extract the support from}
}
\value{
A numeric vector containing the support values.
}
\description{
Returns the support of the input objects
}
