% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{TAMatrix-class}
\alias{TAMatrix-class}
\title{S4 class to represent transactions in a sparse Matrix}
\description{
S4 class to represent transactions in a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{Transaction Matrix in binary sparse representation}
}}

