% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{TAMatrix}
\alias{TAMatrix}
\title{An S4 class to represent an Itemset as a Sparse Matrix}
\description{
An S4 class to represent an Itemset as a Sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{Transaction Matrix in sparse representation}
}}

