% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{vcov.geeglm}
\alias{vcov.geeglm}
\title{estimate the variance-covariance matrix of regression coefficients
from a generalized linear model fitted with generalized estimating
equations.}
\usage{
\method{vcov}{geeglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{geeglm}.}

\item{...}{additional arguments passed along to other functions.}
}
\description{
estimate the variance-covariance matrix of regression coefficients
from a generalized linear model fitted with generalized estimating
equations.
}
