% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AprioriGen.R
\name{AprioriGen}
\alias{AprioriGen}
\title{Apriori Gen of candidates K-1}
\arguments{
\item{L}{Incident matrix of the transactions}
}
\value{
The candidate for the next iteration of Apriori
}
\description{
Takes in a incident matrix of transaction, where the rows do describe the potential items and each colomn does represent the itemsets.
}
