% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationRules.R
\name{AssociationRules}
\alias{AssociationRules}
\title{Calculate Association rules with minimal confidence and support.}
\usage{
AssociationRules(FrequentItemsets, Itemsets, minconfidence, minsupport = NULL,
  arefrequent = TRUE)
}
\arguments{
\item{Itemsets}{This should either be the transaction matrix, the then paramter arefrequent should be set to FALSE or it should be the frequent itemsets with
the paramter arefrequent = TRUE.}

\item{minconfidence}{This should be the minimal confidence the rules are supposed to have.}

\item{minsupport}{Only necessary if the Itemset are not the frequent Itemsets. Then it should describe the minimal support of the rules.}

\item{arefrequent}{Are the input Itemsets the frequent Itemsets ?}
}
\value{
Incident matrix of the association rules with minimal support and confidence.
}
\description{
This function takes either frequent Itemsets as input or calculates them by itself and based on these itemsets it estimates the rules with minimal confidence.
}
