% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineCands.R
\name{CombineCands}
\alias{CombineCands}
\title{Combining incidence matrices of different length}
\usage{
CombineCands(list_input)
}
\arguments{
\item{list_input}{has to be a list of matrices. These matrices should have appropriate rownames describing the different items. The items should represent 
the different candidates.}
}
\value{
The result will be yet another incident matrix that will have all items rows of all input matrices and will represent all itemset in the input matrices.
}
\description{
Combining incidence matrices of different length
}
