% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineCands.R
\name{CombineFIMatrix}
\alias{CombineFIMatrix}
\title{Combining list of FIMatrix containing frequent itemsets with different length.}
\usage{
CombineFIMatrix(list_input)
}
\arguments{
\item{list_input}{List of sparse incident matrices. These matrices may have different number 
of rows, columns and number of items}
}
\value{
Another object of class FIMatrix that contains all the input Itemsets.
}
\description{
This functions takes a list of FIMatrices matrices as input. The FIMatrices may have
 different number or rows (different items), different number of columns 
 (different number of itemsets) as well es different number of items per itemset. Then these
  matrices are combined to a big, sparse incident matrix that does have all rows and columns
   of the input matrices.
}
