% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{CombineRules}
\alias{CombineRules}
\title{Combining list of Rules}
\usage{
CombineRules(list_input)
}
\arguments{
\item{list_input}{List of Rules. These Rules may have different number 
of rows, columns and number of items}
}
\value{
Rules Object that will contain all the rules from the input rules objects.
}
\description{
This function takes as input a list of objects of class Rules that may have different items 
and rules and comines them to a single Rule.
}
