% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineCands.R
\name{CombineRules}
\alias{CombineRules}
\title{Combining list of Rules containing Rules of different length.}
\usage{
CombineRules(list_input)
}
\arguments{
\item{list_input}{List of Rules. These Rules may have different number 
of rows, columns and number of items}
}
\value{
The result will be yet another Rules Object that will contain all the rules from the 
input rules but in one object.
}
\description{
Combining list of Rules containing Rules of different length.
}
