% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{confidence}
\alias{confidence}
\title{Extract confidence from object}
\usage{
confidence(object)
}
\arguments{
\item{object}{Object of Class Rules}
}
\value{
Vector of confidence values from all entities of the objects.
}
\description{
Extract confidence from object
}
