% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{DetRules_1}
\alias{DetRules_1}
\title{Determine all possible association rules with consequent length 1 for given frequent itemsets}
\usage{
DetRules_1(Items)
}
\arguments{
\item{Items}{Object of class FIMatrix containing frequent itemsets of different length. Minimal
length should be two.}
}
\value{
Object of class Rules that contains all relevant rules with consequent length 1.
}
\description{
This function takes frequent itemsets as input and generates all rules of consequent length 1 
that have to be considered for rule mining. It uses the Apriori-gen for more efficient
generation of rule candidates.
}
