% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetRules.R
\name{DetRules_1}
\alias{DetRules_1}
\title{Determine all possible association rules with consequent length 1 for given frequent transaction.}
\usage{
DetRules_1(Items)
}
\arguments{
\item{Items}{Object of class FIItem containing frequent itemsets of different length,
but at least two.}
}
\value{
a list containing the rhs, lhs and support of the candidates. The vector Item_id shows
from which frequent itemset the specific rules comes from and is needed later on to determine 
the frequent itemsets.
}
\description{
This function takes the frequent itemsets as input and generates rules of length 1.
It uses the Apriori-gen for more efficient generation of rule candidates.
}
