% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetRules.R
\name{DetRules_K}
\alias{DetRules_K}
\title{Determine all possible association rules with consequent length 1for given frequent transaction.}
\usage{
DetRules_K(rules)
}
\arguments{
\item{rules}{Object of class Rules from Project_Apriori containing the current rules with 
minimal support and confidence.}
}
\description{
This function take the frequent transactions as input and generates rules of length len. It uses the Apriori-gen for more efficient generation
of rule candidates.
}
