% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.R
\name{Duplicate}
\alias{Duplicate}
\title{Find duplicate columns in a sparse matrix}
\usage{
Duplicate(mat)
}
\arguments{
\item{mat}{a sparse matrix.}
}
\value{
A boolean vector that does have one element for each column in the matrix.
}
\description{
This function is based on a c function from the arules package that finds duplicated columns
in a sparse matrix
}
