% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{Duplicate}
\alias{Duplicate}
\title{Find duplicate columns in a sparse matrix}
\usage{
Duplicate(mat)
}
\arguments{
\item{mat}{A sparse matrix of type ngTMatrix, lgCMatrix or ngCMatrix.}
}
\value{
A boolean vector that does have one element for each column in the matrix. True does
represent a column that is replicated, False means it is not replicated.
}
\description{
This function is based on a c function from the arules package that finds duplicated columns
in a sparse matrix
}
