% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{extract}
\alias{extract}
\title{Extract FIMatrix object from class}
\usage{
extract(object)
}
\arguments{
\item{object}{Object of Class Rules}
}
\value{
Object of Class FIMatrix
}
\description{
Extract FIMatrix object from class
}
