% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractRules.R
\name{ExtractRules}
\alias{ExtractRules}
\title{Extract Rules}
\usage{
ExtractRules(rules, maxNumConsequent = NULL)
}
\arguments{
\item{rules}{The left hand side and the right hand side of the rules incident matrix where the names rows do represent the items. The both matrices should
be saved in a list.}
}
\value{
The rules from the left hand and right hand side in the form of {It1, ... ItN} -> {ITK}.
}
\description{
This function extracts the rule in the form of {It1, ... ItN} -> {ITK}.
}
