% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequentItemsets.R
\name{FindFrequentItemsets}
\alias{FindFrequentItemsets}
\title{Calculate frequent itemset with minimal support}
\usage{
FindFrequentItemsets(dataset, minsupport)
}
\arguments{
\item{dataset}{currently under development}

\item{minsupport}{This defines the minimal level of the support the resulting itemsets should have.}
}
\value{
The is a list that does store the frequent itemsets as a sparse incidence matrix
 as well as the resulting support as a vector.
}
\description{
This function calculate the frequent itemset with a minimal support.
}
