% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{FIMatrix}
\alias{FIMatrix}
\title{An S4 class to represent frequent itemsets in a sparse Matrix}
\description{
An S4 class to represent frequent itemsets in a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{Frequent Itemset Matrix in sparse representation. Rows are the items and the 
columns represent the itemsets.}

\item{\code{support}}{Vector containing the support of all Itemsets}
}}

