% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequentItemsets.R
\name{FrequentItemsets}
\alias{FrequentItemsets}
\title{Frequent Itemsets base on minimal support}
\usage{
FrequentItemsets(dataset, minsupport)
}
\arguments{
\item{dataset}{This should be an incident matrix of the baskets where the rows describe the items and the columns describe the individual bastets. Naming of the items is important}

\item{minsupport}{This defines the minimal level of the support the resulting itemsets should have.}
}
\value{
The is a list that does store the frequent itemsets as a incidence matrix as well as the resulting support as a vector.
}
\description{
This function gives the frequenitemsets given a minimal level of support.
}
