% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCandidates.R
\name{GenCandidates}
\alias{GenCandidates}
\title{Generate Candidate of length k + 1 based on frequent itemsets of length k.}
\usage{
GenCandidates(L)
}
\arguments{
\item{L}{Sparse incident matrix of the candidates.}
}
\value{
The candidates of length k + 1 as a sparse incident matrix. rownames are kept.
}
\description{
Take an sparce, incident matrix with itemsets of length k and generates the candidates of length
k + 1 based on the apriori Candidate Generation. In the matrix the rows should be the items and 
the columns should be the itemssets that represent the candidates.
}
