% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSets.R
\name{GetSets}
\alias{GetSets}
\title{Get frequent itemsets with support from output of frequentItemset object}
\usage{
GetSets(items)
}
\arguments{
\item{items}{Output of the frequentItemsets function}
}
\value{
A dataframe of all the frequent Itemsets with their respective support.
}
\description{
This function takes the ouput of the frequentItemsets function and creates a list of the frequent itemsets with their support from it.
}
