% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{GiveUniqueCol}
\alias{GiveUniqueCol}
\title{Return the unique columns of sparse matrix.}
\usage{
GiveUniqueCol(mat)
}
\arguments{
\item{mat}{Sparse matrix of class ngTMatrix, lgCMatrix or ngCMatrix.}
}
\value{
Returns only non-duplicate columns of matrix. Ordering might be changed.
}
\description{
Return the unique columns of sparse matrix.
}
