% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{hist-FIMatrix}
\alias{hist-FIMatrix}
\alias{hist,FIMatrix-method}
\title{Plot a Histogram of itemset lengths for a FIMatrix object}
\usage{
\S4method{hist}{FIMatrix}(x)
}
\arguments{
\item{x}{Object of class FIMatrix}
}
\value{
Histogram of Itemsize in FIMatrix.
}
\description{
The hist function gives a histogram of the lengths of the itemsets.
}
