% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{leverage}
\alias{leverage}
\title{Extract leverage from object}
\usage{
leverage(object)
}
\arguments{
\item{object}{Object of Class Rules}
}
\value{
Vector of leverage values from all entities of the objects.
}
\description{
Extract leverage from object
}
