% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{lift}
\alias{lift}
\title{Extract lift from object}
\usage{
lift(object)
}
\arguments{
\item{object}{Object of Class Rules}
}
\value{
Vector of lift values from all entities of the objects.
}
\description{
Extract lift from object
}
