% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransactionMatrix.R
\name{makeTransactionMatrix}
\alias{makeTransactionMatrix}
\title{Make Transaction Dataset from arules to sparse matrix}
\usage{
makeTransactionMatrix(mat)
}
\arguments{
\item{mat}{Object of class Itemmatrix form arules}
}
\value{
The non-sparse matrix
}
\description{
Takes a object of class Itemmatrix from arules and make it a  sparse matrix.
}
