% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{plot-FIMatrix}
\alias{plot-FIMatrix}
\alias{plot,FIMatrix-method}
\title{Plot an FI-matrices.}
\usage{
\S4method{plot}{FIMatrix}(x, pch = 1, col = "red")
}
\arguments{
\item{object}{Object of class FIMatrix}
}
\value{
Scatter plot of Itemsize vs support.
}
\description{
The plot function gives a scatter plot with the support on the y-axis and the itemset-length on
the x-axis
}
