% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{plot-FIMatrix}
\alias{plot-FIMatrix}
\alias{plot,FIMatrix-method}
\title{Plot FIMatrix objects}
\usage{
\S4method{plot}{FIMatrix}(x, pch = 1, col = "red")
}
\arguments{
\item{x}{Object of class FIMatrix}

\item{pch}{Size of points that is given to plot function.}

\item{col}{Colour used by plot()}
}
\value{
Scatter plot of itemsize against support
}
\description{
The plot function creates a scatter plot with the support on the y-axis and the itemset-length on
the x-axis
}
