% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{plot-TAMatrix}
\alias{plot-TAMatrix}
\alias{plot,TAMatrix-method}
\title{Plot TAMatrix objects}
\usage{
\S4method{plot}{TAMatrix}(x)
}
\arguments{
\item{x}{Object of class TAMatrix}
}
\description{
The plot function for TAMatrix creates a histogram of the itemset lengths
}
