% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{prune}
\alias{prune}
\title{S4 Generic for pruning FIMatrix, Rules objects}
\usage{
prune(object, ...)
}
\arguments{
\item{input}{Object of class FIMatrix or}

\item{Support}{Minimal support the pruned FIMatrix or Rules should have.}

\item{Confidence}{Minimal confidence the pruned rule should have.}

\item{Lift}{Minimal Lift the pruned rule should have.}

\item{Leverage}{Minimal Leverage the pruned rule should have.}
}
\description{
With this function one can delete all itemsets from an FIMatrix that do not have minimal support
or confidence and all rules from an Rules object that do not have minimal support, confidence,
lift or leverage.
}
