% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{prune}
\alias{prune}
\title{Pruning objects by metric.}
\usage{
prune(object, ...)
}
\arguments{
\item{object}{Objected to be pruned.}

\item{...}{metrics to prune by}
}
\value{
Pruned object of same class as input.
}
\description{
With thes function one can eliminate parts of an object that do not fulfill certain threshholds.
Currently this function is implemented for the Classes FIMarix and Rules.
}
