% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{qplot-FIMatrix}
\alias{qplot-FIMatrix}
\alias{qplot,FIMatrix-method}
\title{Plot an FI-matrices with ggplot}
\usage{
\S4method{qplot}{FIMatrix}(x, col = "red", alpha = 0.1, type = c("hist",
  "scatter"))
}
\arguments{
\item{x}{Object of class FIMatrix}

\item{col}{colour of data points (only scatter plot) per default "red"}

\item{alpha}{alpha value for scatter plot}

\item{type}{character string "hist" or "scatter" depending on wether you want a histogram or scatter plot}
}
\value{
Scatter plot of itemset length against support or histogram of itemset lengths
}
\description{
The qplot function can do a scatter plot with the support on the y-axis and the itemset-length on
the x-axis or a histogram of itemset lengths
}
