% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Rules}
\alias{Rules}
\title{An S4 class to represent an frequent Itemsets in a sparse Matrix}
\description{
An S4 class to represent an frequent Itemsets in a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{lhs}}{Left-hand side}

\item{\code{rhs}}{Right-hande side}

\item{\code{support}}{Support vector of Rules}

\item{\code{confidence}}{Confidence vector of Rules}
}}

