% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{TAMatrix-class}
\alias{TAMatrix-class}
\title{S4 class to represent transactions in a sparse Matrix}
\description{
S4 class to represent transactions in a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{Transactions as Matrix in a ngTMatrix. The rows should be the items and the columns
should represent the itemsets.}

\item{\code{dim}}{A vector of size 2 containing the number of rows and the number of columns.}

\item{\code{items}}{A vector containing the row names / item names of the items.}
}}

