% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{TAMatrix}
\alias{TAMatrix}
\title{An S4 class to represent an itemset as a sparse Matrix}
\description{
An S4 class to represent an itemset as a sparse Matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{Transaction Matrix in binary sparse representation}
}}

